Create fill pattern

Create fill pattern

Usage

create_fill(
  gradientFill = "",
  patternType = "",
  bgColor = NULL,
  fgColor = NULL,
  ...
)

Arguments

gradientFill

complex fills

patternType

various: default is "none", but also "solid", or a color like "gray125"

bgColor

hex8 color with alpha, red, green, blue only for patternFill

fgColor

hex8 color with alpha, red, green, blue only for patternFill

...

...

See also