xlsx reading, writing and editing.

openxlsx simplifies the the process of writing and styling Excel xlsx files from R and removes the dependency on Java.

Details

The openxlsx package uses global options to simplify formatting:

options("openxlsx2.borderColour" = "black")

options("openxlsx2.borderStyle" = "thin")

options("openxlsx2.dateFormat" = "mm/dd/yyyy")

options("openxlsx2.datetimeFormat" = "yyyy-mm-dd hh:mm:ss")

options("openxlsx2.numFmt" = NULL)

options("openxlsx2.paperSize" = 9) ## A4

options("openxlsx2.orientation" = "portrait") ## page orientation

See the Formatting vignette for examples.

Additional options

options("openxlsx2.compressionLevel" = "9") ## set zip compression level, default is "1".

See also