Read from an Excel file or Workbook object

Read data from an Excel file or Workbook object into a data.frame

read_xlsx(
  xlsxFile,
  sheet,
  startRow = 1,
  startCol = NULL,
  rowNames = FALSE,
  colNames = TRUE,
  skipEmptyRows = FALSE,
  skipEmptyCols = FALSE,
  rows = NULL,
  cols = NULL,
  detectDates = TRUE,
  namedRegion,
  na.strings = "#N/A",
  na.numbers = NA,
  check.names = FALSE,
  sep.names = ".",
  fillMergedCells = FALSE,
  ...
)

Arguments

xlsxFile

An xlsx file, Workbook object or URL to xlsx file.

sheet

The name or index of the sheet to read data from.

startRow

first row to begin looking for data.

startCol

first column to begin looking for data.

rowNames

If TRUE, first column of data will be used as row names.

colNames

If TRUE, the first row of data will be used as column names.

skipEmptyRows

If TRUE, empty rows are skipped else empty rows after the first row containing data will return a row of NAs.

skipEmptyCols

If TRUE, empty columns are skipped.

rows

A numeric vector specifying which rows in the Excel file to read. If NULL, all rows are read.

cols

A numeric vector specifying which columns in the Excel file to read. If NULL, all columns are read.

detectDates

If TRUE, attempt to recognize dates and perform conversion.

namedRegion

A named region in the Workbook. If not NULL startRow, rows and cols parameters are ignored.

na.strings

A character vector of strings which are to be interpreted as NA. Blank cells will be returned as NA.

na.numbers

A numeric vector of digits which are to be interpreted as NA. Blank cells will be returned as NA.

check.names

logical. If TRUE then the names of the variables in the data frame are checked to ensure that they are syntactically valid variable names

sep.names

(unimplemented) One character which substitutes blanks in column names. By default, "."

fillMergedCells

If TRUE, the value in a merged cell is given to all cells within the merge.

...

additional arguments passed to wb_to_df()

Value

data.frame

Details

Formulae written using write_formula to a Workbook object will not get picked up by read_xlsx(). This is because only the formula is written and left to be evaluated when the file is opened in Excel. Opening, saving and closing the file with Excel will resolve this.

See also

Examples

xlsxFile <- system.file("extdata", "readTest.xlsx", package = "openxlsx2")
df1 <- read_xlsx(xlsxFile = xlsxFile, sheet = 1, skipEmptyRows = FALSE)
sapply(df1, class)
#>        Var1        Var2        <NA>        Var3        Var4        Var5 
#>   "logical"   "numeric"   "numeric" "character" "character"      "Date" 
#>        Var6        Var7 
#> "character" "character" 

df2 <- read_xlsx(xlsxFile = xlsxFile, sheet = 3, skipEmptyRows = TRUE)
df2$Date <- convert_date(df2$Date)
sapply(df2, class)
#>        Date       value        word        bool      wordZ2 
#>      "Date"   "numeric" "character"   "logical" "character" 
head(df2)
#>          Date     value      word  bool  wordZ2
#> 7  1944-04-26 0.8390764 N-U-B-R-A FALSE FALSE-Z
#> 8  1944-04-25 0.8863800 N-Z-P-S-Y  TRUE  TRUE-Z
#> 9  1944-04-24 0.5741314 C-G-D-X-H  TRUE  TRUE-Z
#> 10 1944-04-23 0.1366065      <NA> FALSE FALSE-Z
#> 11 1944-04-22 0.3692582 B-K-A-O-W  TRUE  TRUE-Z
#> 12 1944-04-21        NA H-P-G-O-K  TRUE  TRUE-Z

df2 <- read_xlsx(
  xlsxFile = xlsxFile, sheet = 3, skipEmptyRows = TRUE,
  detectDates = TRUE
)
sapply(df2, class)
#>        Date       value        word        bool      wordZ2 
#>      "Date"   "numeric" "character"   "logical" "character" 
head(df2)
#>          Date     value      word  bool  wordZ2
#> 7  2014-04-28 0.8390764 N-U-B-R-A FALSE FALSE-Z
#> 8  2014-04-27 0.8863800 N-Z-P-S-Y  TRUE  TRUE-Z
#> 9  2014-04-26 0.5741314 C-G-D-X-H  TRUE  TRUE-Z
#> 10 2014-04-25 0.1366065      <NA> FALSE FALSE-Z
#> 11 2014-04-24 0.3692582 B-K-A-O-W  TRUE  TRUE-Z
#> 12 2014-04-23        NA H-P-G-O-K  TRUE  TRUE-Z

wb <- wb_load(system.file("extdata", "readTest.xlsx", package = "openxlsx2"))
df3 <- read_xlsx(wb, sheet = 2, skipEmptyRows = FALSE, colNames = TRUE)
df4 <- read_xlsx(xlsxFile, sheet = 2, skipEmptyRows = FALSE, colNames = TRUE)
all.equal(df3, df4)
#> [1] TRUE

wb <- wb_load(system.file("extdata", "readTest.xlsx", package = "openxlsx2"))
df3 <- read_xlsx(wb,
  sheet = 2, skipEmptyRows = FALSE,
  cols = c(1, 4), rows = c(1, 3, 4)
)

## URL
##
xlsxFile <- "https://github.com/JanMarvin/openxlsx2/raw/main/inst/extdata/readTest.xlsx"
head(read_xlsx(xlsxFile))
#>    Var1 Var2 NA  Var3  Var4       Var5         Var6    Var7
#> 2  TRUE    1 NA     1     a 2015-02-07 3209324 This #DIV/0!
#> 3  TRUE   NA NA #NUM!     b 2015-02-06         <NA>    <NA>
#> 4  TRUE    2 NA  1.34     c 2015-02-05         <NA>   #NUM!
#> 5 FALSE    2 NA  <NA> #NUM!       <NA>         <NA>    <NA>
#> 6 FALSE    3 NA  1.56     e       <NA>         <NA>    <NA>
#> 7 FALSE    1 NA   1.7     f 2015-02-02         <NA>    <NA>