Replace data cell(s)

Minimal invasive update of cell(s) inside of imported workbooks.

update_cell(
  x,
  wb,
  sheet,
  cell,
  data_class,
  colNames = FALSE,
  removeCellStyle = FALSE,
  na.strings
)

Arguments

x

value you want to insert

wb

the workbook you want to update

sheet

the sheet you want to update

cell

the cell you want to update in Excel connotation e.g. "A1"

data_class

optional data class object

colNames

if TRUE colNames are passed down

removeCellStyle

keep the cell style?

na.strings

optional na.strings argument. if missing #N/A is used. If NULL no cell value is written, if character or numeric this is written (even if NA is part of numeric data)

Examples

   xlsxFile <- system.file("extdata", "update_test.xlsx", package = "openxlsx2")
   wb <- wb_load(xlsxFile)

   # update Cells D4:D6 with 1:3
   wb <- update_cell(x = c(1:3), wb = wb, sheet = "Sheet1", cell = "D4:D6")

   # update Cells B3:D3 (names())
   wb <- update_cell(x = c("x", "y", "z"), wb = wb, sheet = "Sheet1", cell = "B3:D3")

   # update D4 again (single value this time)
   wb <- update_cell(x = 7, wb = wb, sheet = "Sheet1", cell = "D4")

   # add new column on the left of the existing workbook
   wb <- update_cell(x = 7, wb = wb, sheet = "Sheet1", cell = "A4")

   # add new row on the end of the existing workbook
   wb <- update_cell(x = 7, wb = wb, sheet = "Sheet1", cell = "A9")
   wb_to_df(wb)
#>   NA     x   y  z
#> 4  7    NA 2.0  7
#> 5 NA    NA  NA  2
#> 6 NA  TRUE 2.5  3
#> 7 NA FALSE  NA NA
#> 8 NA  TRUE  NA  5
#> 9  7    NA  NA NA