R6 class for a Workbook

R6 class for a Workbook

R6 class for a Workbook

Value

The integer position of the sheet The wbWorkbook object The wbWorkbook object A namedcharacter vector of sheet names in their order. The names represent the original value of the worksheet prior to any character substitutions. The wbWorkbook object The wbWorkbook object The wbWorkbook object The wbWorkbook object The sheet tables. character() if empty The wbWorkbook object The wbWorkbook object The wbWorkbook object The wbWorkbook object The wbWorkbook object The wbWorkbook object Returns sheet visibility The wbWorkbook object The wbWorkbook object

Details

A Workbook

minor helper wrapping xl_open which does the entire same thing

Public fields

sheet_names

sheet_names

calcChain

calcChain

apps

apps

charts

charts

isChartSheet

isChartSheet

connections

connections

Content_Types

Content_Types

app

app

core

core

drawings

drawings

drawings_rels

drawings_rels

drawings_vml

drawings_vml

embeddings

embeddings

externalLinks

externalLinks

externalLinksRels

externalLinksRels

headFoot

headFoot

media

media

metadata

metadata

persons

persons

pivotTables

pivotTables

pivotTables.xml.rels

pivotTables.xml.rels

pivotDefinitions

pivotDefinitions

pivotRecords

pivotRecords

pivotDefinitionsRels

pivotDefinitionsRels

queryTables

queryTables

rowHeights

rowHeights

slicers

slicers

slicerCaches

slicerCaches

sharedStrings

sharedStrings

styles_mgr

styles_mgr

styles_xml

styles_xml

tables

tables

tables.xml.rels

tables.xml.rels

theme

theme

vbaProject

vbaProject

vml

vml

vml_rels

vml_rels

comments

comments

threadComments

threadComments

workbook

workbook

workbook.xml.rels

workbook.xml.rels

worksheets

worksheets

worksheets_rels

worksheets_rels

sheetOrder

The sheet order. Controls ordering for worksheets and worksheet names.

path

path

creator

A character vector of creators

title

title

subject

subject

category

category

datetimeCreated

The datetime (as POSIXt) the workbook is created. Defaults to the current Sys.time() when the workbook object is created, not when the Excel files are saved.

Methods

Public methods

wbWorkbook$new()

wbWorkbook$append()

wbWorkbook$append_sheets()

wbWorkbook$validate_sheet()

wbWorkbook$add_worksheet()

wbWorkbook$clone_worksheet()

wbWorkbook$addChartSheet()

wbWorkbook$add_data()

wbWorkbook$add_data_table()

wbWorkbook$add_formula()

wbWorkbook$save()

wbWorkbook$open()

wbWorkbook$buildTable()

wbWorkbook$createFontNode()

wbWorkbook$get_base_font()

wbWorkbook$set_base_font()

wbWorkbook$get_sheet_names()

wbWorkbook$set_sheet_names()

wbWorkbook$setSheetName()

wbWorkbook$set_row_heights()

wbWorkbook$remove_row_heights()

wbWorkbook$createCols()

wbWorkbook$group_cols()

wbWorkbook$ungroup_cols()

wbWorkbook$remove_col_widths()

wbWorkbook$set_col_widths()

wbWorkbook$group_rows()

wbWorkbook$ungroup_rows()

wbWorkbook$remove_worksheet()

wbWorkbook$add_data_validation()

wbWorkbook$conditional_formatting()

wbWorkbook$merge_cells()

wbWorkbook$unmerge_cells()

wbWorkbook$freeze_pane()

wbWorkbook$add_image()

wbWorkbook$add_plot()

wbWorkbook$print()

wbWorkbook$protect()

wbWorkbook$protect_worksheet()

wbWorkbook$set_creators()

wbWorkbook$add_creators()

wbWorkbook$remove_creators()

wbWorkbook$set_last_modified_by()

wbWorkbook$page_setup()

wbWorkbook$set_header_footer()

wbWorkbook$get_tables()

wbWorkbook$remove_tables()

wbWorkbook$add_filter()

wbWorkbook$remove_filter()

wbWorkbook$grid_lines()

wbWorkbook$add_named_region()

wbWorkbook$remove_named_region()

wbWorkbook$set_order()

wbWorkbook$get_sheet_visibility()

wbWorkbook$set_sheet_visibility()

wbWorkbook$add_page_break()

wbWorkbook$clean_sheet()

wbWorkbook$clone()

Method new()

Creates a new wbWorkbook object

Usage

wbWorkbook$new(
  creator = NULL,
  title = NULL,
  subject = NULL,
  category = NULL,
  datetimeCreated = Sys.time()
)

Arguments

creator

character vector of creators. Duplicated are ignored.

title

title

subject

subject

category

category

datetimeCreated

The datetime (as POSIXt) the workbook is created. Defaults to the current Sys.time() when the workbook object is created, not when the Excel files are saved.

Returns

a wbWorkbook object

Method append()

Append a field. This method is intended for internal use

Usage

wbWorkbook$append(field, value)

Arguments

field

A valid field name

value

A value for the field

Method append_sheets()

Append to self$workbook$sheets This method is intended for internal use

Usage

wbWorkbook$append_sheets(value)

Arguments

value

A value for self$workbook$sheets

Method validate_sheet()

validate sheet

Usage

wbWorkbook$validate_sheet(sheet)

Arguments

sheet

A character sheet name or integer location

Method add_worksheet()

Add worksheet to the wbWorkbook object

Usage

wbWorkbook$add_worksheet(
  sheet,
  gridLines = TRUE,
  tabColour = NULL,
  zoom = 100,
  header = NULL,
  footer = NULL,
  oddHeader = header,
  oddFooter = footer,
  evenHeader = header,
  evenFooter = footer,
  firstHeader = header,
  firstFooter = footer,
  visible = c("true", "false", "hidden", "visible", "veryhidden"),
  hasDrawing = FALSE,
  paperSize = getOption("openxlsx2.paperSize", default = 9),
  orientation = getOption("openxlsx2.orientation", default = "portrait"),
  hdpi = getOption("openxlsx2.hdpi", default = getOption("openxlsx2.dpi", default =
    300)),
  vdpi = getOption("openxlsx2.vdpi", default = getOption("openxlsx2.dpi", default =
    300))
)

Arguments

sheet

sheet

gridLines

gridLines

tabColour

tabColour

zoom

zoom

header

header

footer

footer

oddHeader

oddHeader

oddFooter

oddFooter

evenHeader

evenHeader

evenFooter

evenFooter

firstHeader

firstHeader

firstFooter

firstFooter

visible

visible

hasDrawing

hasDrawing

paperSize

paperSize

orientation

orientation

hdpi

hdpi

vdpi

vdpi

Returns

The wbWorkbook object, invisibly

Method clone_worksheet()

Clone a workbooksheet

Usage

wbWorkbook$clone_worksheet(old, new)

Arguments

old

name of worksheet to clone

new

name of new worksheet to add

Method addChartSheet()

Add a chart sheet to the workbook

Usage

wbWorkbook$addChartSheet(sheet, tabColour = NULL, zoom = 100)

Arguments

sheet

sheet

tabColour

tabColour

zoom

zoom

Returns

The wbWorkbook object, invisibly

Method add_data()

add data

Usage

wbWorkbook$add_data(
  sheet,
  x,
  startCol = 1,
  startRow = 1,
  array = FALSE,
  xy = NULL,
  colNames = TRUE,
  rowNames = FALSE,
  withFilter = FALSE,
  name = NULL,
  sep = ", ",
  removeCellStyle = TRUE
)

Arguments

sheet

sheet

x

x

startCol

startCol

startRow

startRow

array

array

xy

xy

colNames

colNames

rowNames

rowNames

withFilter

withFilter

name

name

sep

sep

removeCellStyle

removeCellStyle

return

The wbWorkbook object

Method add_data_table()

add a data table

Usage

wbWorkbook$add_data_table(
  sheet,
  x,
  startCol = 1,
  startRow = 1,
  xy = NULL,
  colNames = TRUE,
  rowNames = FALSE,
  tableStyle = "TableStyleLight9",
  tableName = NULL,
  withFilter = TRUE,
  sep = ", ",
  firstColumn = FALSE,
  lastColumn = FALSE,
  bandedRows = TRUE,
  bandedCols = FALSE
)

Arguments

sheet

sheet

x

x

startCol

startCol

startRow

startRow

xy

xy

colNames

colNames

rowNames

rowNames

tableStyle

tableStyle

tableName

tableName

withFilter

withFilter

sep

sep

firstColumn

firstColumn

lastColumn

lastColumn

bandedRows

bandedRows

bandedCols

bandedCols

Method add_formula()

add formula

Usage

wbWorkbook$add_formula(
  sheet,
  x,
  startCol = 1,
  startRow = 1,
  array = FALSE,
  xy = NULL
)

Arguments

sheet

sheet

x

x

startCol

startCol

startRow

startRow

array

array

xy

xy

Method save()

Save the workbook

Usage

wbWorkbook$save(path = self$path, overwrite = TRUE)

Arguments

path

The path to save the workbook to

overwrite

If FALSE, will not overwrite when path exists

Returns

The wbWorkbook object invisibly

Method open()

open wbWorkbook in Excel.

Usage

wbWorkbook$open(interactive = NA)

Arguments

interactive

If FALSE will throw a warning and not open the path. This can be manually set to TRUE, otherwise when NA (default) uses the value returned from base::interactive()

Returns

The wbWorkbook, invisibly

Method buildTable()

Build table

Usage

wbWorkbook$buildTable(
  sheet,
  colNames,
  ref,
  showColNames,
  tableStyle,
  tableName,
  withFilter,
  totalsRowCount = 0,
  showFirstColumn = 0,
  showLastColumn = 0,
  showRowStripes = 1,
  showColumnStripes = 0
)

Arguments

sheet

sheet

colNames

colNames

ref

ref

showColNames

showColNames

tableStyle

tableStyle

tableName

tableName

withFilter

withFilter

totalsRowCount

totalsRowCount

showFirstColumn

showFirstColumn

showLastColumn

showLastColumn

showRowStripes

showRowStripes

showColumnStripes

showColumnStripes

Returns

The wbWorksheet object, invisibly

Method createFontNode()

Create a font node from a style

Usage

wbWorkbook$createFontNode(style)

Arguments

style

style

Returns

The font node as xml?

Method get_base_font()

Get the base font

Usage

wbWorkbook$get_base_font()

Returns

A list of of the font

Method set_base_font()

Get the base font

Usage

wbWorkbook$set_base_font(
  fontSize = 11,
  fontColour = "black",
  fontName = "Calibri"
)

Arguments

fontSize

fontSize

fontColour

fontColour

fontName

fontName

Returns

The wbWorkbook object

Method get_sheet_names()

Get sheet names

Usage

wbWorkbook$get_sheet_names()

Method set_sheet_names()

Sets a sheet name

Usage

wbWorkbook$set_sheet_names(old = NULL, new)

Arguments

old

Old sheet name

new

New sheet name

Returns

The wbWorkbook object, invisibly

Method setSheetName()

Deprecated. Use set_sheet_names() instead

Usage

wbWorkbook$setSheetName(sheet, name)

Arguments

sheet

Old sheet name

name

New sheet name

Returns

The wbWorkbook object, invisibly

Method set_row_heights()

Sets a row height for a sheet

Usage

wbWorkbook$set_row_heights(sheet, rows, heights)

Arguments

sheet

sheet

rows

rows

heights

heights

Returns

The wbWorkbook object, invisibly

Method remove_row_heights()

Sets a row height for a sheet

Usage

wbWorkbook$remove_row_heights(sheet, rows)

Arguments

sheet

sheet

rows

rows

Returns

The wbWorkbook object, invisibly description creates column object for worksheet

Method createCols()

Usage

wbWorkbook$createCols(sheet, n, beg, end)

Arguments

sheet

sheet

n

n

beg

beg

end

end

Method group_cols()

Group cols

Usage

wbWorkbook$group_cols(sheet, cols, collapsed = FALSE, levels = NULL)

Arguments

sheet

sheet

cols

cols

collapsed

collapsed

levels

levels

Returns

The wbWorkbook object, invisibly

Method ungroup_cols()

ungroup cols

Usage

wbWorkbook$ungroup_cols(sheet, cols)

Arguments

sheet

sheet

cols

= cols

Method remove_col_widths()

Remove row heights from a worksheet

Usage

wbWorkbook$remove_col_widths(sheet, cols)

Arguments

sheet

A name or index of a worksheet

cols

Indices of columns to remove custom width (if any) from.

Returns

The wbWorkbook object, invisibly

Method set_col_widths()

Group cols

Usage

wbWorkbook$set_col_widths(sheet, cols, widths = 8.43, hidden = FALSE)

Arguments

sheet

sheet

cols

cols

widths

Width of columns

hidden

A logical vector to determine which cols are hidden; values are repeated across length of cols

Returns

The wbWorkbook object, invisibly

Method group_rows()

Group rows

Usage

wbWorkbook$group_rows(sheet, rows, collapsed = FALSE, levels = NULL)

Arguments

sheet

sheet

rows

rows

collapsed

collapsed

levels

levels

Returns

The wbWorkbook object, invisibly

Method ungroup_rows()

ungroup rows

Usage

wbWorkbook$ungroup_rows(sheet, rows)

Arguments

sheet

sheet

rows

rows

Returns

The wbWorkbook object

Method remove_worksheet()

Remove a worksheet

Usage

wbWorkbook$remove_worksheet(sheet)

Arguments

sheet

The worksheet to delete

Returns

The wbWorkbook object, invisibly

Method add_data_validation()

Adds data validation

Usage

wbWorkbook$add_data_validation(
  sheet,
  cols,
  rows,
  type,
  operator,
  value,
  allowBlank = TRUE,
  showInputMsg = TRUE,
  showErrorMsg = TRUE
)

Arguments

sheet

sheet

cols

cols

rows

rows

type

type

operator

operator

value

value

allowBlank

allowBlank

showInputMsg

showInputMsg

showErrorMsg

showErrorMsg

Method conditional_formatting()

Set conditional formatting for a sheet

Usage

wbWorkbook$conditional_formatting(
  sheet,
  startRow,
  endRow,
  startCol,
  endCol,
  dxfId,
  formula,
  type,
  values,
  params
)

Arguments

sheet

sheet

startRow

startRow

endRow

endRow

startCol

startCol

endCol

endCol

dxfId

dxfId

formula

formula

type

type

values

values

params

params

Returns

The wbWorkbook object, invisibly

Method merge_cells()

Set cell merging for a sheet

Usage

wbWorkbook$merge_cells(sheet, rows = NULL, cols = NULL)

Arguments

sheet

sheet

rows, cols

Row and column specifications.

Returns

The wbWorkbook object, invisibly

Method unmerge_cells()

Removes cell merging for a sheet

Usage

wbWorkbook$unmerge_cells(sheet, rows = NULL, cols = NULL)

Arguments

sheet

sheet

rows, cols

Row and column specifications.

Returns

The wbWorkbook object, invisibly

Method freeze_pane()

Set freeze panes for a sheet

Usage

wbWorkbook$freeze_pane(
  sheet,
  firstActiveRow = NULL,
  firstActiveCol = NULL,
  firstRow = FALSE,
  firstCol = FALSE
)

Arguments

sheet

sheet

firstActiveRow

firstActiveRow

firstActiveCol

firstActiveCol

firstRow

firstRow

firstCol

firstCol

Returns

The wbWorkbook object, invisibly

Method add_image()

Insert an image into a sheet

Usage

wbWorkbook$add_image(
  sheet,
  file,
  width = 6,
  height = 3,
  startRow = 1,
  startCol = 1,
  rowOffset = 0,
  colOffset = 0,
  units = "in",
  dpi = 300
)

Arguments

sheet

sheet

file

file

width

width

height

height

startRow

startRow

startCol

startCol

rowOffset

rowOffset

colOffset

colOffset

units

units

dpi

dpi

Returns

The wbWorkbook object, invisibly

Method add_plot()

Add plot. A wrapper for add_image()

Usage

wbWorkbook$add_plot(
  sheet,
  width = 6,
  height = 4,
  xy = NULL,
  startRow = 1,
  startCol = 1,
  rowOffset = 0,
  colOffset = 0,
  fileType = "png",
  units = "in",
  dpi = 300
)

Arguments

sheet

sheet

width

width

height

height

xy

xy

startRow

startRow

startCol

startCol

rowOffset

rowOffset

colOffset

colOffset

fileType

fileType

units

units

dpi

dpi

Method print()

Prints the wbWorkbook object

Usage

wbWorkbook$print()

Returns

The wbWorkbook object, invisibly; called for its side-effects

Method protect()

Protect a workbook

Usage

wbWorkbook$protect(
  protect = TRUE,
  password = NULL,
  lockStructure = FALSE,
  lockWindows = FALSE,
  type = c("1", "2", "4", "8"),
  fileSharing = FALSE,
  username = unname(Sys.info()["user"]),
  readOnlyRecommended = FALSE
)

Arguments

protect

protect

password

password

lockStructure

lockStructure

lockWindows

lockWindows

type

type

fileSharing

fileSharing

username

username

readOnlyRecommended

readOnlyRecommended

Returns

The wbWorkbook object, invisibly

Method protect_worksheet()

protect worksheet

Usage

wbWorkbook$protect_worksheet(
  sheet,
  protect = TRUE,
  password = NULL,
  properties = NULL
)

Arguments

sheet

sheet

protect

protect

password

password

properties

A character vector of properties to lock. Can be one or more of the following: "selectLockedCells", "selectUnlockedCells", "formatCells", "formatColumns", "formatRows", "insertColumns", "insertRows", "insertHyperlinks", "deleteColumns", "deleteRows", "sort", "autoFilter", "pivotTables", "objects", "scenarios"

Method set_creators()

Set creator(s)

Usage

wbWorkbook$set_creators(creators)

Arguments

creators

A character vector of creators to set. Duplicates are ignored.

Method add_creators()

Add creator(s)

Usage

wbWorkbook$add_creators(creators)

Arguments

creators

A character vector of creators to add. Duplicates are ignored.

Method remove_creators()

Remove creator(s)

Usage

wbWorkbook$remove_creators(creators)

Arguments

creators

A character vector of creators to remove. All duplicated are removed.

Method set_last_modified_by()

Change the last modified by

Usage

wbWorkbook$set_last_modified_by(LastModifiedBy = NULL)

Arguments

LastModifiedBy

A new value

Returns

The wbWorkbook object, invisibly

Method page_setup()

page_setup()

Usage

wbWorkbook$page_setup(
  sheet,
  orientation = NULL,
  scale = 100,
  left = 0.7,
  right = 0.7,
  top = 0.75,
  bottom = 0.75,
  header = 0.3,
  footer = 0.3,
  fitToWidth = FALSE,
  fitToHeight = FALSE,
  paperSize = NULL,
  printTitleRows = NULL,
  printTitleCols = NULL,
  summaryRow = NULL,
  summaryCol = NULL
)

Arguments

sheet

sheet

orientation

orientation

scale

scale

left

left

right

right

top

top

bottom

bottom

header

header

footer

footer

fitToWidth

fitToWidth

fitToHeight

fitToHeight

paperSize

paperSize

printTitleRows

printTitleRows

printTitleCols

printTitleCols

summaryRow

summaryRow

summaryCol

summaryCol

Returns

The wbWorkbook object, invisibly

Method set_header_footer()

Sets headers and footers

Usage

wbWorkbook$set_header_footer(
  sheet,
  header = NULL,
  footer = NULL,
  evenHeader = NULL,
  evenFooter = NULL,
  firstHeader = NULL,
  firstFooter = NULL
)

Arguments

sheet

sheet

header

header

footer

footer

evenHeader

evenHeader

evenFooter

evenFooter

firstHeader

firstHeader

firstFooter

firstFooter

Returns

The wbWorkbook object, invisibly

Method get_tables()

get tables

Usage

wbWorkbook$get_tables(sheet)

Arguments

sheet

sheet

Method remove_tables()

remove tables

Usage

wbWorkbook$remove_tables(sheet, table)

Arguments

sheet

sheet

table

table

Method add_filter()

add filters

Usage

wbWorkbook$add_filter(sheet, rows, cols)

Arguments

sheet

sheet

rows

rows

cols

cols

Method remove_filter()

remove filters

Usage

wbWorkbook$remove_filter(sheet)

Arguments

sheet

sheet

Method grid_lines()

grid lines

Usage

wbWorkbook$grid_lines(sheet, show = FALSE)

Arguments

sheet

sheet

show

show

Method add_named_region()

add a named region

Usage

wbWorkbook$add_named_region(
  sheet,
  cols,
  rows,
  name,
  localSheetId = NULL,
  overwrite = FALSE
)

Arguments

sheet

sheet

cols

cols

rows

rows

name

name

localSheetId

localSheetId

overwrite

overwrite

Method remove_named_region()

remove a named region

Usage

wbWorkbook$remove_named_region(sheet = NULL, name = NULL)

Arguments

sheet

sheet

name

name

Method set_order()

set worksheet order

Usage

wbWorkbook$set_order(sheets)

Arguments

sheets

sheets

Returns

The wbWorkbook object

Method get_sheet_visibility()

Get sheet visibility

Usage

wbWorkbook$get_sheet_visibility()

Method set_sheet_visibility()

Set sheet visibility

Usage

wbWorkbook$set_sheet_visibility(sheet, value)

Arguments

sheet

sheet

value

value

Method add_page_break()

Add a page break

Usage

wbWorkbook$add_page_break(sheet, row = NULL, col = NULL)

Arguments

sheet

sheet

row

row

col

col

Method clean_sheet()

clean sheet (remove all values)

Usage

wbWorkbook$clean_sheet(
  sheet,
  numbers = TRUE,
  characters = TRUE,
  styles = TRUE,
  merged_cells = TRUE
)

Arguments

sheet

sheet

numbers

remove all numbers

characters

remove all characters

styles

remove all styles

merged_cells

remove all merged_cells

Returns

The wbWorksheetObject, invisibly Deep cloning method for workbooks. This method also accesses $clone(deep = TRUE) methods for R6 fields.

Method clone()

The objects of this class are cloneable with this method.

Usage

wbWorkbook$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.