add font for cell region

add font for cell region

wb_add_font(
  wb,
  sheet,
  dims,
  name = "Calibri",
  color = c(rgb = "FF000000"),
  size = "11",
  bold = "",
  italic = "",
  outline = "",
  strike = "",
  underline = "",
  charset = "",
  condense = "",
  extend = "",
  family = "",
  scheme = "",
  shadow = "",
  vertAlign = ""
)

Arguments

wb

a workbook

sheet

the worksheet

dims

the cell range

name

font name: default "Calibri"

color

rgb color: default "FF000000"

size

font size: default "11",

bold

bold, "single" or "double", default: ""

italic

italic

outline

outline

strike

strike

underline

underline

charset

charset

condense

condense

extend

extend

family

font family

scheme

font scheme

shadow

shadow

vertAlign

vertical alignment

Value

The wbWorksheetObject, invisibly

Details

add_font provides all the options openxml accepts for a font node, not all have to be set. Usually name, size and color should be what the user wants.

Examples

 wb <- wb_workbook() %>% wb_add_worksheet("S1") %>% wb_add_data("S1", mtcars)
 wb %>% wb_add_font("S1", "A1:K1", name = "Arial", color = c(theme = "4"))
#> A Workbook object.
#>  
#> Worksheets:
#>  
#>  Worksheet write order: 1