Add a character vector as an Excel Formula

Add a character vector containing Excel formula to a worksheet.

Usage

wb_add_formula(
  wb,
  sheet = current_sheet(),
  x,
  dims = wb_dims(start_row, start_col),
  start_col = 1,
  start_row = 1,
  array = FALSE,
  cm = FALSE,
  apply_cell_style = TRUE,
  remove_cell_style = FALSE,
  ...
)

Arguments

wb

A Workbook object containing a worksheet.

sheet

The worksheet to write to. Can be the worksheet index or name.

x

A character vector.

dims

Spreadsheet dimensions that will determine startCol and startRow: "A1", "A1:B2", "A:B"

start_col

A vector specifying the starting column to write to.

start_row

A vector specifying the starting row to write to.

array

A bool if the function written is of type array

cm

A special kind of array function that hides the curly braces in the cell. Add this, if you see "@" inserted into your formulas

apply_cell_style

Should we write cell styles to the workbook

remove_cell_style

keep the cell style?,

...

additional arguments

Details

Currently only the English version of functions are supported. Please don't use the local translation. The examples below show a small list of possible formulas:

SUM(B2:B4)

AVERAGE(B2:B4)

MIN(B2:B4)

MAX(B2:B4)

...

See also