Add pivot table to a worksheet

add pivot table

Usage

wb_add_pivot_table(
  wb,
  x,
  sheet = next_sheet(),
  dims = "A3",
  filter,
  rows,
  cols,
  data,
  fun,
  params
)

Arguments

wb

A Workbook object containing a #' worksheet.

x

a wb_data object

sheet

a worksheet

dims

the worksheet cell where the pivot table is placed

filter

a character object with names used to filter

rows

a character object with names used as rows

cols

a character object with names used as cols

data

a character object with names used as data

fun

a character object of functions to be used with the data

params

a list of parameters to modify pivot table creation

Details

fun can be either of AVERAGE, COUNT, COUNTA, MAX, MIN, PRODUCT, STDEV, STDEVP, SUM, VAR, VARP.

The sheet will be empty unless it is opened in spreadsheet software.

See also

Examples

wb <- wb_workbook() %>% wb_add_worksheet() %>% wb_add_data(x = mtcars)

df <- wb_data(wb)

wb <- wb %>%
wb_add_pivot_table(df, dims = "A3",
  filter = "am", rows = "cyl", cols = "gear", data = "disp")